refactor(site): extract useChatStream hook from AgentDetail
Move all WebSocket/SSE streaming state management from AgentDetail.tsx
into a dedicated useChatStream hook:

- 6 useState calls (messagesById, streamState, streamError,
  queuedMessages, chatStatus, subagentStatusOverrides)
- streamResetFrameRef and related scheduling callbacks
- updateSidebarChat callback (only used by SSE handler)
- 3 sync effects (chatMessages, chatRecord, chatQueuedMessages)
- SSE event handler useEffect (172 lines)

AgentDetail.tsx reduced from 693 to 434 lines. The hook exposes a
clean interface via UseChatStreamOptions/UseChatStreamResult,
keeping the component focused on queries, mutations, and rendering.
