+++ site/src/pages/AgentsPage/AgentDetail.tsx
@@ -1,35 +1,21 @@
-import { API, watchChat } from "api/api";
+import { API } from "api/api";
 import {
 	chat,
-	chatDiffContentsKey,
 	chatDiffStatus,
-	chatDiffStatusKey,
-	chatKey,
 	chatModels,
 	chats,
-	chatsKey,
 	createChatMessage,
 	deleteChatQueuedMessage,
 	interruptChat,
 	promoteChatQueuedMessage,
 } from "api/queries/chats";
 import type * as TypesGen from "api/typesGenerated";
-import { asRecord, asString } from "components/ai-elements/runtimeTypeUtils";
 import { displayError } from "components/GlobalSnackbar/utils";
 import { Skeleton } from "components/Skeleton/Skeleton";
 import { getVSCodeHref, SESSION_TOKEN_PLACEHOLDER } from "modules/apps/apps";
-import {
-	type FC,
-	startTransition,
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	useState,
-} from "react";
+import { type FC, useEffect, useMemo, useRef, useState } from "react";
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import { useNavigate, useOutletContext, useParams } from "react-router";
-import type { OneWayMessageEvent } from "utils/OneWayWebSocket";
 import { AgentChatInput } from "./AgentChatInput";
 import { ConversationTimeline } from "./AgentDetail/ConversationTimeline";
 import {
@@ -43,12 +29,9 @@ import {
 	buildSubagentTitles,
 	parseMessagesWithMergedTools,
 } from "./AgentDetail/messageParsing";
-import {
-	applyMessagePartToStreamState,
-	buildStreamTools,
-} from "./AgentDetail/streamState";
+import { buildStreamTools } from "./AgentDetail/streamState";
 import { AgentDetailTopBarPortals } from "./AgentDetail/TopBarPortals";
-import type { StreamState } from "./AgentDetail/types";
+import { useChatStream } from "./AgentDetail/useChatStream";
 import { useMessageWindow } from "./AgentDetail/useMessageWindow";
 import type { AgentsOutletContext } from "./AgentsPage";
 import {
@@ -77,20 +60,6 @@ export const AgentDetail: FC = () => {
 	const { agentId } = useParams<{ agentId: string }>();
 	const outletContext = useOutletContext<AgentsOutletContext | undefined>();
 	const queryClient = useQueryClient();
-	const [messagesById, setMessagesById] = useState<
-		Map<number, TypesGen.ChatMessage>
-	>(new Map());
-	const [streamState, setStreamState] = useState<StreamState | null>(null);
-	const [streamError, setStreamError] = useState<string | null>(null);
-	const [queuedMessages, setQueuedMessages] = useState<
-		readonly TypesGen.ChatQueuedMessage[]
-	>([]);
-	const [chatStatus, setChatStatus] = useState<TypesGen.ChatStatus | null>(
-		null,
-	);
-	const [subagentStatusOverrides, setSubagentStatusOverrides] = useState<
-		Map<string, TypesGen.ChatStatus>
-	>(new Map());
 	const [selectedModel, setSelectedModel] = useState("");
 	const [showDiffPanel, setShowDiffPanel] = useState(false);
 	const chatErrorReasons = outletContext?.chatErrorReasons ?? {};
@@ -102,7 +71,6 @@ export const AgentDetail: FC = () => {
 		outletContext?.setRightPanelOpen ?? noopSetRightPanelOpen;
 	const requestArchiveAgent =
 		outletContext?.requestArchiveAgent ?? noopRequestArchiveAgent;
-	const streamResetFrameRef = useRef<number | null>(null);
 	const scrollContainerRef = useRef<HTMLDivElement | null>(null);
 
 	const chatQuery = useQuery({
@@ -162,76 +130,23 @@ export const AgentDetail: FC = () => {
 		promoteChatQueuedMessage(queryClient, agentId ?? ""),
 	);
 
-	const updateSidebarChat = useCallback(
-		(updater: (chat: TypesGen.Chat) => TypesGen.Chat) => {
-			if (!agentId) {
-				return;
-			}
-
-			queryClient.setQueryData<readonly TypesGen.Chat[] | undefined>(
-				chatsKey,
-				(currentChats) => {
-					if (!currentChats) {
-						return currentChats;
-					}
-
-					let didUpdate = false;
-					const nextChats = currentChats.map((chat) => {
-						if (chat.id !== agentId) {
-							return chat;
-						}
-						didUpdate = true;
-						return updater(chat);
-					});
-
-					return didUpdate ? nextChats : currentChats;
-				},
-			);
-		},
-		[agentId, queryClient],
-	);
-
-	const cancelScheduledStreamReset = useCallback(() => {
-		if (streamResetFrameRef.current === null) {
-			return;
-		}
-		window.cancelAnimationFrame(streamResetFrameRef.current);
-		streamResetFrameRef.current = null;
-	}, []);
-
-	const scheduleStreamReset = useCallback(() => {
-		cancelScheduledStreamReset();
-		streamResetFrameRef.current = window.requestAnimationFrame(() => {
-			setStreamState(null);
-			streamResetFrameRef.current = null;
-		});
-	}, [cancelScheduledStreamReset]);
-
-	useEffect(() => {
-		if (!chatMessages) {
-			setMessagesById(new Map());
-			return;
-		}
-		setMessagesById(
-			new Map(chatMessages.map((message) => [message.id, message])),
-		);
-	}, [chatMessages]);
-
-	useEffect(() => {
-		if (!chatRecord) {
-			setChatStatus(null);
-			return;
-		}
-		setChatStatus(chatRecord.status);
-	}, [chatRecord]);
-
-	useEffect(() => {
-		if (!chatData) {
-			setQueuedMessages([]);
-			return;
-		}
-		setQueuedMessages(chatQueuedMessages ?? []);
-	}, [chatData, chatQueuedMessages]);
+	const {
+		messagesById,
+		streamState,
+		chatStatus,
+		streamError,
+		queuedMessages,
+		subagentStatusOverrides,
+		clearStreamError,
+	} = useChatStream({
+		chatId: agentId,
+		chatMessages,
+		chatRecord,
+		chatData,
+		chatQueuedMessages,
+		setChatErrorReason,
+		clearChatErrorReason,
+	});
 
 	useEffect(() => {
 		if (!chatModelConfig) {
@@ -244,179 +159,6 @@ export const AgentDetail: FC = () => {
 			return resolveModelFromChatConfig(chatModelConfig, modelOptions);
 		});
 	}, [chatModelConfig, modelOptions]);
-
-	useEffect(() => {
-		if (!agentId) {
-			return;
-		}
-
-		cancelScheduledStreamReset();
-		setStreamState(null);
-		setStreamError(null);
-		setSubagentStatusOverrides(new Map());
-
-		const socket = watchChat(agentId);
-		const handleMessage = (
-			payload: OneWayMessageEvent<TypesGen.ServerSentEvent>,
-		) => {
-			if (payload.parseError || !payload.parsedMessage) {
-				setStreamError("Failed to parse chat stream update.");
-				return;
-			}
-			if (payload.parsedMessage.type !== "data") {
-				return;
-			}
-
-			const streamEvent = payload.parsedMessage
-				.data as TypesGen.ChatStreamEvent & Record<string, unknown>;
-			if (!streamEvent?.type) {
-				return;
-			}
-
-			switch (streamEvent.type) {
-				case "message": {
-					const message = streamEvent.message;
-					if (!message) {
-						return;
-					}
-					setMessagesById((prev) => {
-						const next = new Map(prev);
-						next.set(message.id, message);
-						return next;
-					});
-					scheduleStreamReset();
-					updateSidebarChat((chat) => ({
-						...chat,
-						updated_at: message.created_at ?? new Date().toISOString(),
-					}));
-					void queryClient.invalidateQueries({ queryKey: chatsKey });
-					return;
-				}
-				case "message_part": {
-					const part = asRecord(streamEvent.message_part?.part);
-					if (!part) {
-						return;
-					}
-					cancelScheduledStreamReset();
-					startTransition(() => {
-						setStreamState((prev) => applyMessagePartToStreamState(prev, part));
-					});
-					return;
-				}
-				case "queue_update": {
-					const queuedMsgs = streamEvent.queued_messages;
-					setQueuedMessages(queuedMsgs ?? []);
-					return;
-				}
-				case "status": {
-					const status = asRecord(streamEvent.status);
-					const nextStatus = asString(status?.status) as TypesGen.ChatStatus;
-					if (!nextStatus) {
-						return;
-					}
-
-					const eventChatID = asString(streamEvent.chat_id);
-					if (eventChatID && eventChatID !== agentId) {
-						setSubagentStatusOverrides((prev) => {
-							if (prev.get(eventChatID) === nextStatus) {
-								return prev;
-							}
-							const next = new Map(prev);
-							next.set(eventChatID, nextStatus);
-							return next;
-						});
-						return;
-					}
-
-					setChatStatus(nextStatus);
-					if (agentId && nextStatus !== "error") {
-						clearChatErrorReason(agentId);
-					}
-					updateSidebarChat((chat) => ({
-						...chat,
-						status: nextStatus,
-						updated_at: new Date().toISOString(),
-					}));
-					if (agentId) {
-						void Promise.all([
-							queryClient.invalidateQueries({
-								queryKey: chatDiffStatusKey(agentId),
-							}),
-							queryClient.invalidateQueries({
-								queryKey: chatDiffContentsKey(agentId),
-							}),
-						]);
-					}
-					const shouldRefreshQueries =
-						nextStatus === "completed" ||
-						nextStatus === "error" ||
-						nextStatus === "paused" ||
-						nextStatus === "waiting";
-					if (shouldRefreshQueries) {
-						void Promise.all([
-							queryClient.invalidateQueries({ queryKey: chatsKey }),
-							queryClient.invalidateQueries({
-								queryKey: chatKey(agentId),
-							}),
-						]);
-					}
-					return;
-				}
-				case "error": {
-					const error = asRecord(streamEvent.error);
-					const reason =
-						asString(error?.message).trim() || "Chat processing failed.";
-					setChatStatus("error");
-					setStreamError(reason);
-					if (agentId) {
-						setChatErrorReason(agentId, reason);
-					}
-					updateSidebarChat((chat) => ({
-						...chat,
-						status: "error",
-						updated_at: new Date().toISOString(),
-					}));
-					void Promise.all([
-						queryClient.invalidateQueries({ queryKey: chatsKey }),
-						queryClient.invalidateQueries({
-							queryKey: chatKey(agentId),
-						}),
-					]);
-					return;
-				}
-				default:
-					break;
-			}
-		};
-
-		const handleError = () => {
-			setStreamError((current) => current ?? "Chat stream disconnected.");
-			void Promise.all([
-				queryClient.invalidateQueries({ queryKey: chatsKey }),
-				queryClient.invalidateQueries({
-					queryKey: chatKey(agentId),
-				}),
-			]);
-		};
-
-		socket.addEventListener("message", handleMessage);
-		socket.addEventListener("error", handleError);
-
-		return () => {
-			socket.removeEventListener("message", handleMessage);
-			socket.removeEventListener("error", handleError);
-			socket.close();
-			cancelScheduledStreamReset();
-		};
-	}, [
-		agentId,
-		cancelScheduledStreamReset,
-		clearChatErrorReason,
-		queryClient,
-		scheduleStreamReset,
-		setChatErrorReason,
-		updateSidebarChat,
-	]);
 
 	const messages = useMemo(() => {
 		const list = Array.from(messagesById.values());
@@ -473,7 +215,7 @@ export const AgentDetail: FC = () => {
 			model: selectedModel || undefined,
 		};
 		clearChatErrorReason(agentId);
-		setStreamError(null);
+		clearStreamError();
 		if (scrollContainerRef.current) {
 			scrollContainerRef.current.scrollTop = 0;
 		}
